+++ src/SynExtend.h
@@ -28,3 +28,7 @@ SEXP calcDValue(SEXP tnPtr, SEXP occVec);
 SEXP calcDRandValue(SEXP tnPtr, SEXP allLabels, SEXP numP, SEXP iterNum);
 SEXP calcDBrownValue(SEXP tnPtr, SEXP allLabels, SEXP iterNum, SEXP SD, SEXP START, SEXP THRESH);
 
+/**** XORRand.c ****/
+SEXP pseudoRandomSample(SEXP N);
+SEXP randomProjection(SEXP VEC, SEXP NONZERO, SEXP N, SEXP OUTDIM);
+SEXP seededPseudoRandomSample(SEXP N, SEXP SEED);